video-mixer: Fix the way we release mixer pads
We were using the actual mixer pad to release the smart mixer
pad, which seemed to be on purpose, but was not properly handle,
moreover, it is now forbiden to pass a pad not inside a GstElement
when releasing it.

Also properly remove ghost pads from Smart mixer, we were planly
failling at it.